% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{cnum}
\alias{cnum}
\title{cnum  to see column numbers (used within ncomma.pmisc and nplus.pmisc)}
\usage{
cnum(x)
}
\arguments{
\item{x}{}
}
\description{
cnum  to see column numbers (used within ncomma.pmisc and nplus.pmisc)
}
