% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{ncomma}
\alias{ncomma}
\title{ncomma() returns show names(df) in alphabetical order and in multiple format}
\usage{
ncomma(x)
}
\arguments{
\item{x}{}
}
\description{
ncomma() returns show names(df) in alphabetical order and in multiple format
}
\examples{
\dontrun{
mtcars \%>\% dplyr::select(starts_with("c")  ) \%>\% names \%>\% ncomma

}

data.frame: ncomma() returns show names(df) in alphabetical order and in multiple format
character: ncomma() returns vector in multiple forms to facilitate cut-and-paste (Yikes! for experts; Yay for PYH)
mtcars \%>\% ncomma
mtcars \%>\% dplyr::select(starts_with("c")  ) \%>\% names \%>\% ncomma
}
