% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{nfactor}
\alias{nfactor}
\title{nfactor used in conjunction with fct_relevel to manually arrange factor levels}
\usage{
nfactor(x)
}
\arguments{
\item{x}{}
}
\description{
nfactor used in conjunction with fct_relevel to manually arrange factor levels
}
\examples{
\dontrun{

 used in conjunction with fct_relevel to manually arrange factor levels

}
}
