% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhestci.R
\name{yhestci}
\alias{yhestci}
\title{yhestci returns vector with estimate and CI pasted together
inspired by https://rdrr.io/github/malcolmbarrett/mbmisc/src/R/formatrs.R}
\usage{
yhestci(
  mydf,
  estimate,
  lcl,
  ucl = NULL,
  divider = c(" to "),
  est_digit = 2,
  ci_digit = 2
)
}
\arguments{
\item{ci_digit}{}
}
\description{
yhestci returns vector with estimate and CI pasted together
inspired by https://rdrr.io/github/malcolmbarrett/mbmisc/src/R/formatrs.R
}
\examples{
\dontrun{
mtcars \%>\% data.frame() \%>\% mutate(myestci = yhestci(mydf=mtcars, "hp", "cyl", "mpg", est_digit = 0))
est/ci_digits determined by another variable (e.g., cyl)
mtcars \%>\% group_by(am) \%>\% nest() \%>\%
mutate(myestci = map(data, ~ yhestci(mydf=.x, "hp", "wt", "mpg", est_digit = 0, ci_digit = .x$cyl[[1]] ))) \%>\%  unnest(myestci)
}




yhestci returns vector with estimate and CI pasted together
sometimes does not play nice with tibble + group_by()
yhestci() supersedes yhcorprint()
mtcars \%>\% data.frame() \%>\% mutate(myestci = yhestci(mydf=mtcars, "hp", "cyl", "mpg", est_digit = 0))
est/ci_digits determined by another variable (e.g., cyl)
mtcars \%>\% group_by(am) \%>\% nest() \%>\%
mutate(myestci = map(data, ~ yhestci(mydf=.x, "hp", "wt", "mpg", est_digit = 0, ci_digit = .x$cyl[[1]] ))) \%>\%  unnest(myestci)
}
