% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{yhfixid}
\alias{yhfixid}
\title{clean up inconsistent ptid}
\usage{
yhfixid(idvar, width = 3, name_prefix = "ID")
}
\arguments{
\item{idvar}{assume idvar is coded as ID001}

\item{name_prefix}{}
}
\description{
clean up inconsistent ptid
}
\examples{
\dontrun{
mtcars \%>\% mutate(carb_id = yhfixid(carb, name_prefix = "mtcars"))
}
clean up inconsistent ptid
assume idvar is coded as ID001
mtcars \%>\% mutate(carb_id = yhfixid(carb, name_prefix = "mtcars"))
}
