% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhnmsd.R
\name{yhnmsd}
\alias{yhnmsd}
\title{Descriptive statistics of response variable by group and timepoint}
\usage{
yhnmsd(data, mydv, dtpoint, tgroup, est_d = 2, ci_d = 2, propcut)
}
\arguments{
\item{data}{(dataframe) dataframe of the study}

\item{mydv}{(string)    response variable}

\item{dtpoint}{(string)    discrete timepoints}

\item{tgroup}{(string)    treatment group variable}

\item{est_d}{(numeric)   number of decimal places for point estimate}

\item{ci_d}{(numeric)   number of decimal places for CI}

\item{propcut}{(numeric)   cutpoint to create proportions (descriptive purposes)}
}
\value{
a dataframe
}
\description{
Descriptive statistics of response variable by group and timepoint
}
