% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhprop.R
\name{yhprop}
\alias{yhprop}
\title{compute prop with Wilson 95pct for continuous and ordinal variables}
\usage{
yhprop(y1)
}
\arguments{
\item{y1}{}
}
\description{
compute prop with Wilson 95pct for continuous and ordinal variables
}
\examples{
\dontrun{
yhprop (mtcars$hp >200)
yhprop (mtcars$cyl>=6)
}


}
