% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhnmsd.R
\name{yhround_sub}
\alias{yhround_sub}
\title{round variables}
\usage{
yhround_sub(x, digits, trail = TRUE)
}
\arguments{
\item{x}{variable}

\item{digits}{number of digits}

\item{trail}{trailing zeros}
}
\description{
round variables
}
