% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhnmsd.R
\name{yhnmsd}
\alias{yhnmsd}
\title{Number, mean (or proportion) SD (or percentages)}
\usage{
yhnmsd(data, mydv, dtpoint, tgroup, est_d = 2, ci_d = 2, propcut)
}
\arguments{
\item{data}{(dataframe) dataframe of the study}

\item{mydv}{(string)    response variable}

\item{dtpoint}{(string)    discrete timepoints}

\item{tgroup}{(string)    treatment group variable}

\item{est_d}{(numeric)   number of decimal places for point estimate}

\item{ci_d}{(numeric)   number of decimal places for CI}

\item{propcut}{(numeric)   cutpoint to create proportions (descriptive purposes)}
}
\value{
a dataframe
}
\description{
This function provides descriptive statistics of response variable by study groups and discrete timepoints.
I'm aware that such functions are "gao-dua-du" (Hokkien: "ubiquitous") in CRAN but I have customized this for my needs.
}
\examples{
\dontrun{
yhnmsd(data = mtcars, mydv = "qsec", dtpoint = "am", tgroup = "vs")
}
}
